add logging and error checking id

+++ src/controllers/TenantController.ts
@@ -3,6 +3,7 @@ import { TenantService } from "../services/tenantService";
 import { TenantRequest } from "../types";
 import { Logger } from "winston";
 import { validationResult } from "express-validator";
+import createHttpError from "http-errors";
 export class TenantController {
   constructor(
     private tenantService: TenantService,
@@ -39,6 +40,12 @@ export class TenantController {
   }
   async getById(req: Request, res: Response, next: NextFunction) {
     const { id } = req.params;
+    if (isNaN(Number(id))) {
+      const error = createHttpError(400, "Invalid url param.");
+      next(error);
+      return;
+    }
+    this.logger.debug("Request to fetch tenant by id", { id });
     try {
       const tenant = await this.tenantService.getById(Number(id));
       res.json(tenant);
@@ -49,8 +56,16 @@ export class TenantController {
   }
   async deleteById(req: Request, res: Response, next: NextFunction) {
     const { id } = req.params;
+    if (isNaN(Number(id))) {
+      const error = createHttpError(400, "Invalid url param.");
+      next(error);
+      return;
+    }
+
+    this.logger.debug("Request to delete tenant", { id });
     try {
       await this.tenantService.deleteById(Number(id));
+      this.logger.info("Tenant delete successful", { id });
       res.json();
     } catch (err) {
       next(err);
@@ -64,10 +79,16 @@ export class TenantController {
     }
 
     const { id } = req.params;
+    if (isNaN(Number(id))) {
+      const error = createHttpError(400, "Invalid url param.");
+      next(error);
+      return;
+    }
     const { name, address } = req.body;
-
+    this.logger.debug("Request to update tenant", { id, name, address });
     try {
       await this.tenantService.updateById(Number(id), { name, address });
+      this.logger.info("Tenant update successful", { id });
       res.json();
     } catch (err) {
       next(err);
